测试在hash下用scaling的效果

+++ gp_nerf/sample_bg.py
@@ -12,7 +12,7 @@ def  bg_sample_inv(near, far, point_num, device):
 # NOTE: 2023/08/08  不在网络部分使用scaling， 改为在contract时使用
 def contract_to_unisphere(x: torch.Tensor, hparams):
 
-    z_range = torch.tensor(hparams.z_range)
+    z_range = torch.tensor(hparams.z_range, dtype=torch.float32)
     aabb = torch.tensor([z_range[0], hparams.sphere_center[1] - hparams.sphere_radius[1], hparams.sphere_center[2] - hparams.sphere_radius[2], 
                          z_range[1], hparams.sphere_center[1] + hparams.sphere_radius[1], hparams.sphere_center[2] + hparams.sphere_radius[2]]).to(x.device)
